+++ commands/eco/eco.js
@@ -1,57 +1,61 @@
 module.exports = {
   name: "eco",
-  category: "eco",
+  category: "botowner",
   description: "Adjust a user's economy balance.",
-  usage: "<GIVE | SET | TAKE> <@USER> <VALUE>",
+  usage: "<ADD | SUBTRACT | SET> <@USER> <VALUE>",
   example: "set @Steve#1234 69000",
   permission: "BOT OWNER",
   run: async (bot, message, args) => {
 
-    const Discord = require("discord.js");
-    if (!message.guild) return;
-    const db = require("quick.db");
+      const Discord = require("discord.js");
+      const ecoUtils = require("../../main/functions/ecoUtils");
 
-    if (message.author.id !== bot.settings.ids.botOwner)
-      return;
+      if (message.author.id !== bot.settings.ids.botOwner) return;
 
-    let subc = args[0];
-    let user = message.mentions.members.first();
-    let amount = args[2];
+      let person = message.mentions.members.first();
+      if (!person) return message.reply("Please specify a user");
 
-    if (subc === "give") {
-      if (!user)
-        return message.reply(`Please specify a user to give money to.`);
-      if (!amount)
-        return message.reply(
-          `Please specify how much money you would like to give this person.`
-        );
+      let amount = args[2];
+      if (!amount) return message.reply("Please specify an amount.");
 
-      message.reply("Successfully added " + amount + " to " + user);
-      db.add(`money_${user.id}`, amount);
-    } else if (subc === "take") {
-      if (!user)
-        return message.reply(`Please specify a user to take money from.`);
-      if (!amount)
-        return message.reply(
-          `Please specify how much money you would like to take from this person.`
-        );
+      switch (args[0]) {
+        
+          case "add":
+              await ecoUtils.getUser(person.id).then(async (user) => {
+                  let IntAmount = parseInt(amount);
+                  let value = user.balance + IntAmount;
+                  await ecoUtils.updateUser(person.id, value).then((user) => {
+                      return bot.createEmbed("info", "", `${amount} has now been added to ${person}'s balance. Their new balance is ${user.balance}.`, [], ``, bot)
+                          .then((embed) => message.channel.send(embed))
+                          .catch((error) => bot.logger("error", error));
+                  })
+              })
+              break;
 
-      message.reply(`Successfully removed ${amount} from ${user}'s account.`);
-      db.subtract(`money_${user.id}`, amount);
-    } else if (subc === "set") {
-      if (!user) return message.reply(`Please specify a user to set money to.`);
-      if (!amount)
-        return message.reply(
-          `Please specify how much money you would like to set this person's balance to.`
-        );
+          case "subtract":
+              await ecoUtils.getUser(person.id).then(async (user) => {
+                let IntAmount = parseInt(amount);
+                let value = user.balance - IntAmount;
+                  if (value < 0) return message.reply("I cannot do this as the user will go into a negative balance.");
+                  await ecoUtils.updateUser(person.id, value).then((user) => {
+                      return bot.createEmbed("info", "", `${amount} has now been subtracted from ${person}'s balance. Their new balance is ${user.balance}.`, [], ``, bot)
+                          .then((embed) => message.channel.send(embed))
+                          .catch((error) => bot.logger("error", error));
+                  })
+              })
+              break;
 
-      message.reply(
-        "This option currently doesn't work at the moment. Please use the Give or Take money command."
-      );
-    } else if (!subc || subc == "help") {
-        return bot.helpEmbed("eco", bot)
-        .then((embed) => message.channel.send(embed))
-        .catch((error) => bot.logger("error", error));
-    }
-  },
+          case "set":
+              await ecoUtils.updateUser(person.id, amount).then((user) => {
+                  return bot.createEmbed("info", "", `${person}'s balance has now been set to ${user.balance}.`, [], ``, bot)
+                      .then((embed) => message.channel.send(embed))
+                      .catch((error) => bot.logger("error", error));
+              })
+              break;
+
+          default:
+              message.reply("Provide a subcommand. `add`, `subtract` or `set`");
+              break;
+      }
+  }
 };